[1,2,3 단계] Web MVC (#5)
* [step1] 1단계 요구사항 추가

* [step1] 토큰 API 생성 및 토큰 값 검증 추가

* [step1] 토큰 발급 기능 완성

* [step1] 코드 정리

* [step1] jwt 패키지 생성

* [step1] 토큰을 이용한 내 정보 조회하기 기능 구현

* [step2] 2단계 요구사항 추가

* [step2] Member에 role 추가

* [step2] AuthMember 와 ArgumentResolver 추가

* [step2] 로그인 유저만 예약 가능하도록 수정

* [step2] 자신의 예약이 아닌 경우 예약 취소하도록 추가

* [step2] 패키지 변경

* [step2] 자신의 예약인지 비교하는 로직 수정

* [step3] 3단계 요구사항 추가

* [step3] 관리자 기능 API에 /admin 추가

* [step3] 애플리케이션 실행시 관리자 추가

* [step3] 관리자 API를 인터셉터를 이용하여 인증 인가를 진행한다.

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,16 +1,19 @@
 package nextstep.reservation;
 
+import java.util.List;
+import java.util.Objects;
+import nextstep.auth.AuthMember;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
+import nextstep.support.AuthorizationException;
 import nextstep.support.DuplicateEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
 @Service
 public class ReservationService {
+
     public final ReservationDao reservationDao;
     public final ThemeDao themeDao;
     public final ScheduleDao scheduleDao;
@@ -21,7 +24,7 @@ public class ReservationService {
         this.scheduleDao = scheduleDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, AuthMember authMember) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
             throw new NullPointerException();
@@ -33,8 +36,9 @@ public class ReservationService {
         }
 
         Reservation newReservation = new Reservation(
-                schedule,
-                reservationRequest.getName()
+            schedule,
+            reservationRequest.getName(),
+            authMember.getId()
         );
 
         return reservationDao.save(newReservation);
@@ -49,12 +53,14 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, AuthMember authMember) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
         }
-
+        if (!Objects.equals(authMember.getId(), reservation.getMemberId())) {
+            throw new AuthorizationException();
+        }
         reservationDao.deleteById(id);
     }
 }
